ignore bleu score in doctest, renamed to nlp.py
