+++ pytorch_lightning/metrics/functional/__init__.py
@@ -27,4 +27,4 @@ from pytorch_lightning.metrics.functional.regression import (
     rmse,
     rmsle,
 )
-from pytorch_lightning.metrics.functional.sequence import bleu_score
+from pytorch_lightning.metrics.functional.nlp import bleu_score
